fix: workaround for FF's absence of an svg.onload event
Firefox does not not emit a `load` event on `<svg>` elements when
they're appended to the DOM and reference no external file (I assume
they only emit `load` when the element actually references an external
file). To get around this, simply execute the "on load" behavior just
after the element is added to the DOM.

The reason for the onload hook in the first place was so that
`svg.clientHeight` and `svg.clientWidth` would accurately reflect the
parent's height/width, which might not be so until the element is fully
emplaced in the DOM. On Firefox, however, these values are updated
immediately and so there's no need to put this initialization logic in a
separate onload hook.